requirements.txt
chainlit==1.1.202
faiss-cpu==1.8.0
fastapi==0.110.3
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.1
langchain-experimental==0.0.59
langchain-google-genai
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langdetect==1.0.9
langserve==0.2.0
langsmith==0.1.63
lxml
openai
open_clip_torch
pillow
pydantic
tiktoken
torch
unstructured[pdf]
uvicorn==0.25.0